--- src/Routes/admin/AdminRoutes.ts
@@ -1,34 +1,0 @@
-import express from "express";
-
-import multer from "multer";
-import path from "path";
-import fs from "fs";
-
-const {
-  verifyToken,
-} = require("../../Controller/Authentication/AuthenticationControllers");
-
-const {
-  UserCountController,
-  UserlistController,
-  CheckAPIController
-} = require("../../Controller/admin/adminController");
-
-const AdminRoutes = express.Router();
-
-AdminRoutes.get("/userCount",
-    verifyToken,
- UserCountController);
-
-AdminRoutes.post("/userList",
-    verifyToken,
- UserlistController
-);
-
-AdminRoutes.get("/CheckAPI",
-    // verifyToken,
- CheckAPIController
-);
-
-
-export default AdminRoutes;
